[Feat]#52-controller,userdata수정

+++ back/src/main/java/com/rookiefit/back/service/implement/UserDataServiceImplement.java
@@ -3,11 +3,10 @@ package com.rookiefit.back.service.implement;
 import java.io.IOException;
 import java.util.List;
 
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import com.rookiefit.back.dto.request.userData.GetUserBodyDataRequestDto;
-import com.rookiefit.back.dto.request.userData.GetUserProfileRequestDto;
 import com.rookiefit.back.dto.request.userData.InputUserBodyDataRequestDto;
 import com.rookiefit.back.dto.request.userData.InputUserProfileRequestDto;
 import com.rookiefit.back.dto.response.userData.GetUserBodyDataResponseDto;
@@ -15,10 +14,11 @@ import com.rookiefit.back.dto.response.userData.GetUserProfileResponseDto;
 import com.rookiefit.back.dto.response.userData.InputUserBodyDataResponseDto;
 import com.rookiefit.back.dto.response.userData.InputUserProfileResponseDto;
 import com.rookiefit.back.entity.UserBodyDataEntity;
+import com.rookiefit.back.entity.UserEntity;
 import com.rookiefit.back.entity.UserProfileEntity;
-import com.rookiefit.back.provider.JwtProvider;
 import com.rookiefit.back.repository.UserBodyDataRepository;
 import com.rookiefit.back.repository.UserProfileRepository;
+import com.rookiefit.back.repository.UserRepository;
 import com.rookiefit.back.service.FirebaseService;
 import com.rookiefit.back.service.UserDataService;
 
@@ -29,69 +29,60 @@ import lombok.RequiredArgsConstructor;
 public class UserDataServiceImplement implements UserDataService {
 
         private final UserProfileRepository userProfileRepository;
+        private final UserRepository userRepository;
         private final UserBodyDataRepository userBodyDataRepository;
         private final FirebaseService firebaseService;
-        private final JwtProvider jwtProvider;
 
         @Override
-        public ResponseEntity<? super InputUserProfileResponseDto> inputUserProfile(InputUserProfileRequestDto dto) {
-                String currentUserId = jwtProvider.getUserIdFromToken(dto.getToken());
-
+        public ResponseEntity<? super InputUserProfileResponseDto> inputUserProfile(InputUserProfileRequestDto dto, String currentUserId) {
                 boolean isExsitedId = userProfileRepository.existsByUserAuthEntity_UserId(currentUserId);
-                if (isExsitedId) {
-                        userProfileRepository.deleteAllByUserAuthEntity_UserId(currentUserId);
-                }
-                dto.setToken(currentUserId);
-
+                if (isExsitedId) userProfileRepository.deleteAllByUserAuthEntity_UserId(currentUserId);
                 String uploadedFileUrl = null;
                 if (dto.getUserProfileImageFile() != null && !dto.getUserProfileImageFile().isEmpty()) {
                         try {
                         uploadedFileUrl = firebaseService.uploadFile(dto.getUserProfileImageFile());
                         } catch (IOException exception) {
-                         exception.printStackTrace();
+                                exception.printStackTrace();
                         }
                 }
-                UserProfileEntity userProfileEntity = new UserProfileEntity(dto,uploadedFileUrl);
+                if(dto.getUserNickname() == null){ 
+                        dto.setUserNickname("닉네임없는 헬린이");
+                        System.out.println(dto.getUserNickname());
+                }
+                UserEntity userEntity = userRepository.findByUserId(currentUserId);
+                UserProfileEntity userProfileEntity = new UserProfileEntity(dto,uploadedFileUrl,userEntity);
                 userProfileRepository.save(userProfileEntity);
                 return InputUserProfileResponseDto.succes();
         }
 
         @Override
-        public ResponseEntity<? super GetUserProfileResponseDto> getUserProfile(GetUserProfileRequestDto dto) {
-                String currentUserId = jwtProvider.getUserIdFromToken(dto.getToken());
+        public ResponseEntity<? super GetUserProfileResponseDto> getUserProfile(String currentUserId) {
                 boolean isExsitedId = userProfileRepository.existsByUserAuthEntity_UserId(currentUserId);
-                if (!isExsitedId) {
-                        System.out.println("아이디 존재하지 않음");
-                }
+                if (!isExsitedId) return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null); 
 
                 UserProfileEntity userProfileEntity = userProfileRepository.findByUserAuthEntity_UserId(currentUserId);
                 return GetUserProfileResponseDto.success(userProfileEntity);
         }
 
+        //241204-19:18_김민준 리팩토링
         @Override
-        public ResponseEntity<? super InputUserBodyDataResponseDto> inputUserBodyData(InputUserBodyDataRequestDto dto) {
-                String currentId = jwtProvider.getUserIdFromToken(dto.getToken());
+        public ResponseEntity<? super InputUserBodyDataResponseDto> inputUserBodyData(InputUserBodyDataRequestDto dto, String currentUserId) {
                 String currentDate = dto.getInbodydate();
                 boolean isExsitedId = userBodyDataRepository.existsByInbodydate(currentDate);
-                if (isExsitedId) {
-                        userBodyDataRepository.deleteAllByInbodydate(currentDate);
-                }
-                dto.setToken(currentId);
+                if (isExsitedId)userBodyDataRepository.deleteAllByInbodydate(currentDate);
                 // UserProfileEntity 조회 (현재 userId를 기반으로 UserProfileEntity를 찾기)
-                UserProfileEntity userProfileEntity = userProfileRepository.findByUserAuthEntity_UserId(currentId);
-                UserBodyDataEntity userBodyDataEntity = new UserBodyDataEntity(dto);
-                userBodyDataEntity.setUserProfileData(userProfileEntity);
+                UserProfileEntity userProfileEntity = userProfileRepository.findByUserAuthEntity_UserId(currentUserId);
+                UserBodyDataEntity userBodyDataEntity = new UserBodyDataEntity(dto,userProfileEntity);
 
                 userBodyDataRepository.save(userBodyDataEntity);
                 return InputUserBodyDataResponseDto.success();
         }
 
         @Override
-        public ResponseEntity<List<GetUserBodyDataResponseDto>> getUserBodyData(GetUserBodyDataRequestDto dto) {
-                String currentUserId = jwtProvider.getUserIdFromToken(dto.getToken());
+        public ResponseEntity<List<GetUserBodyDataResponseDto>> getUserBodyData(String currentUserId) {
                 boolean isExsitedId = userBodyDataRepository.existsByUserProfile_UserAuthEntity_UserId(currentUserId);
                 if (!isExsitedId) {
-                        System.out.println("아이디가존재하지않음");
+                        return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
                 }
                 List<UserBodyDataEntity> responseDtos = userBodyDataRepository.findUserBodyDataByUserId(currentUserId);
                 return GetUserBodyDataResponseDto.success(responseDtos);
